Do not import Danger globals in the dangerfile
See https://github.com/danger/danger-js/blob/master/source/danger.ts for
the scoop. There's a throw there, so evaluating the dangerfile fails
when running danger locally with `danger ci` or `danger pr`. This should
be harmless because the dangerfile is (and should) only ever be imported
by danger itself, which injects the `danger` global.